requirements.txt
streamlit
langchain
langchain-community
filelock
uuid
python-dotenv
sentence-transformers
chromadb
PyPDF2
openai
